requirements.txt
SQLAlchemy~=2.0.40
fastapi~=0.115.12
psycopg2-binary~=2.9.10
python-dotenv
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
python-multipart~=0.0.6
redis~=5.0.1